requirements.txt
alpaca-py>=0.30.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
python-dotenv>=1.0.0
schedule>=1.2.0
requests>=2.31.0
tabulate>=0.9.0
colorama>=0.4.6
flask>=2.2.0
flask-cors>=4.0.0
